restructured graphql type definitions and added README.md
