--- src/typeDefs/chats/index.ts
@@ -1,2 +1,0 @@
-export * from "./chat";
-export * from "./message";
