--- src/typeDefs/donation.ts
@@ -1,13 +1,0 @@
-import { gql } from "apollo-server-core";
-
-export const donation = gql`
-  type Donation {
-    _id: ID!
-    userId: ID!
-    orgId: ID!
-    payPalId: String!
-    nameOfUser: String!
-    nameOfOrg: String!
-    amount: Float!
-  }
-`;
