+++ src/typeDefs/interfaces.ts
@@ -1,0 +1,4 @@
+import { gql } from "apollo-server-core";
+
+// Place fields alphabetically to ensure easier lookup and navigation.
+export const interfaces = gql``;
