--- src/typeDefs/plugin/index.ts
@@ -1,2 +1,0 @@
-export * from "./plugin";
-export * from "./pluginField";
