--- src/typeDefs/plugin/pluginField.ts
@@ -1,10 +1,0 @@
-import { gql } from "apollo-server-core";
-
-export const pluginField = gql`
-  type PluginField {
-    key: String!
-    value: String!
-    status: Status!
-    createdAt: String
-  }
-`;
