+++ src/typeDefs/scalars.ts
@@ -1,0 +1,6 @@
+import { gql } from "apollo-server-core";
+
+// Place fields alphabetically to ensure easier lookup and navigation.
+export const scalars = gql`
+  scalar Upload
+`;
